requirements.txt
PyQt5==5.15.11
mysql-connector-python==9.5.0